Booking Methods
Added almost all the booking methods along with test templates for them.
These templates need to be replaced in future.

All that's left to do is to wrap the actual booking creation (inserting
the row and incrementing the statistics) in a transaction and call them
in the appropriate router method.